backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
Pillow==10.1.0
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy<2.0.0
scikit-learn
aiofiles==23.2.0
pydantic==2.5.0